% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\docType{data}
\name{data_categoryname}
\alias{data_categoryname}
\title{Category Name Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_categoryname
}
\description{
A data frame with the available category name data for the Yahoo Finance API.
}
\keyword{datasets}
