% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\docType{data}
\name{data_errors}
\alias{data_errors}
\title{Errors Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_errors
}
\description{
A data frame with the available errors data for the Yahoo Finance API.
}
\keyword{datasets}
