% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\docType{data}
\name{data_exchange}
\alias{data_exchange}
\title{Exchange Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_exchange
}
\description{
A data frame with the available exchange data for the Yahoo Finance API.
}
\keyword{datasets}
