% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\docType{data}
\name{data_filters}
\alias{data_filters}
\title{Filters Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_filters
}
\description{
A data frame with the available filters data for the Yahoo Finance API.
}
\keyword{datasets}
