% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\docType{data}
\name{data_fundfamilyname}
\alias{data_fundfamilyname}
\title{Fund Family Name Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_fundfamilyname
}
\description{
A data frame with the available fund family name data for the Yahoo Finance API.
}
\keyword{datasets}
