% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\docType{data}
\name{data_industry}
\alias{data_industry}
\title{Industry Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_industry
}
\description{
A data frame with the available industry name for the Yahoo Finance API.
}
\keyword{datasets}
