% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\docType{data}
\name{data_peer_group}
\alias{data_peer_group}
\title{Peer Group Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_peer_group
}
\description{
A data frame with the available peer group data for the Yahoo Finance API.
}
\keyword{datasets}
