% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\docType{data}
\name{data_sector}
\alias{data_sector}
\title{Sector Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_sector
}
\description{
A data frame with the available sector data for the Yahoo Finance API.
}
\keyword{datasets}
