% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene_plot.R
\name{DE_gene_plot}
\alias{DE_gene_plot}
\title{function definitions ##### Plot DE gene numbers of each perturbations}
\usage{
DE_gene_plot(
  score_dir,
  pval_dir,
  project = "perturb",
  prefix = "./",
  label = "",
  p_val_cut = 0.05,
  score_cut = 0.5,
  ylimit = c(-600, 600, 200)
)
}
\description{
function definitions ##### Plot DE gene numbers of each perturbations
}
