% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{get_crumb_and_cookies}
\alias{get_crumb_and_cookies}
\title{Get the Crumb, Cookies, and Handle for Yahoo Finance API}
\usage{
get_crumb_and_cookies()
}
\value{
A list containing the following elements:
\item{handle}{A curl handle object to be used for subsequent requests.}
\item{crumb}{A string representing the crumb value for authentication.}
\item{cookies}{A data frame of cookies retrieved during the request.}
}
\description{
A function to get the crumb, cookies, and handle required to authenticate and interact
with the Yahoo Finance API.
}
\examples{
session <- get_crumb_and_cookies()
}
