% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{get_screen}
\alias{get_screen}
\title{Get Screen Data from the Yahoo Finance API}
\usage{
get_screen(payload = create_payload())
}
\arguments{
\item{payload}{list. Payload to send to the Yahoo Finance API created using
the \code{\link{create_query}} and \code{\link{create_payload}} functions.}
}
\value{
A data frame containing data from the Yahoo Finance API for the specified screen.
}
\description{
A function to send a payload to the Yahoo Finance API and get data for the screen.
}
\examples{
filters <- list(
  list("or", list(
    list("eq", list("region", "us"))
  ))
)

query <- create_query(filters)

payload <- create_payload(
  quote_type = "equity", query = query,
  size = 25, offset = 0,
  sort_field = "intradaymarketcap", sort_type = "desc",
  top_operator = "and"
)

screen <- get_screen(payload)
}
