% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegratedMixscape.R
\name{IntegratedMixscape}
\alias{IntegratedMixscape}
\title{function definitions ##### Integrated some functions in Seurat as mentioned in
vignette of Mixscape.Prefer Seurat object after QC.}
\usage{
IntegratedMixscape(sg_dir, mtx_dir, NTC = "NTC", replicate = 1, prefix = "./")
}
\description{
function definitions ##### Integrated some functions in Seurat as mentioned in
vignette of Mixscape.Prefer Seurat object after QC.
}
