% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\docType{data}
\name{data_region}
\alias{data_region}
\title{Regions Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_region
}
\description{
A data frame with the available regions for the Yahoo Finance API.
}
\keyword{datasets}
