% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCREEN.R
\name{SCREEN}
\alias{SCREEN}
\title{function definitions ##### Run all function with one command}
\usage{
SCREEN(
  sg_dir,
  mtx_dir,
  species = "Hs",
  prefix = "./",
  label = "",
  gene_frac = 0.01,
  nFeature = c(200, 5000),
  nCount = 1000,
  mt = 10,
  blank_NTC = FALSE,
  lambda = 0.01,
  permutation = NULL,
  p_val_cut = 0.05,
  score_cut = 0.5,
  ylimit = c(-600, 600, 200),
  project = "perturb",
  NTC = "NTC",
  replicate = 1,
  select_gene = NULL
)
}
\description{
function definitions ##### Run all function with one command
}
